pyproject.toml
[project]
name = "vyapar-gst-env"
version = "0.1.0"
description = "OpenEnv environment for Indian GST compliance — derived from the Vyapar SME assistant."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
